Changed both DAQs to compile off same files
Compressing code a little by having DAQ2 compile off of original DAQ.cpp and config.h files. Behavior is defined by build/compile flags in pio -- similar to the system used for actuator controllers.
