+++ platformio.ini
@@ -50,11 +50,11 @@ build_flags = -DHAL=teensy -DETH
 
 [env:DAQ]
 src_filter = -<*> +<DAQ/> -<DAQ/DAQ2.cpp>
-build_flags = -DHAL=teensy -DETH
+build_flags = -DHAL=teensy -DDAQ1 -DETH
 
 [env:DAQ2]
-src_filter = -<*> +<DAQ/> -<DAQ/DAQ.cpp>
-build_flags = -DHAL=teensy -DETH
+src_filter = -<*> +<DAQ/> -<DAQ/DAQ2.cpp>
+build_flags = -DHAL=teensy -DDAQ2 -DETH
 
 [env:DAQ_spoof]
 src_filter = -<*> +<DAQ/>
