+++ src/DAQ/config.h
@@ -1,5 +1,6 @@
 #include <Analog_Thermocouple.h>
 #include <Cryo_Thermocouple.h>
+#include <SPI_Thermocouple.h>
 #include <common_fw.h>
 #include <ADS8167.h>
 #include <Automation.h>
@@ -11,6 +12,7 @@
 
 #define FLIGHT_BRAIN_ADDR 0x00
 
+#ifdef DAQ1
 #ifdef ETH
 IPAddress ip(10, 0, 0, 11); // dependent on local network
 #endif
@@ -56,13 +58,34 @@ uint8_t battMonINAAddr = 0x40;
 const uint8_t numSensors = 5;
 sensorInfo sensors[numSensors];
 
+#elif DAQ2
+#ifdef ETH
+IPAddress ip(10, 0, 0, 12); // dependent on local network
+#endif
+const uint8_t numCryoTherms = 4;
+// therm[2] = lox adapter tree pt, therm[3] = lox adapter tree gems
+// ADDR = GND, VDD, 10k & 4.3K, 10K & 13K
+uint8_t cryoThermCS[numCryoTherms] = {21, 20, 19, 18};
+uint8_t cryoThermCLK = 13;
+uint8_t cryoThermDO = 12; 
+Adafruit_MAX31855 _cryo_boards[numCryoTherms];
+float cryoReadsBackingStore[numCryoTherms];
+
+const uint8_t numSensors = 2;
+sensorInfo sensors[numSensors];
+
+#endif
+
 const float batteryMonitorShuntR = 0.002; // ohms
 const float batteryMonitorMaxExpectedCurrent = 10; // amps
 
 const float powerSupplyMonitorShuntR = 0.010; // ohms
 const float powerSupplyMonitorMaxExpectedCurrent = 5; // amps
 
+
+
 namespace config {
+  #ifdef DAQ1
   void setup() {
     debug("Initializing ADCs");
     for (int i = 0; i < numADCSensors; i++) {
@@ -110,6 +133,12 @@ namespace config {
     sensors[3] = {"Number Packets Sent", FLIGHT_BRAIN_ADDR, 5, 10};
     sensors[4] = {"Analog Thermocouples", FLIGHT_BRAIN_ADDR, 19, 3};
     // sensors[6] = {"Power Supply Stats", FLIGHT_BRAIN_ADDR, 6, 3};
-
   }
+  #elif DAQ2
+  void setup() {
+    debug("Initializing sensors");
+    sensors[0] = {"Cryo Temps",      FLIGHT_BRAIN_ADDR, 4, 3};
+    sensors[1] = {"Number Packets Sent", FLIGHT_BRAIN_ADDR, 5, 10};
+  }
+  #endif 
 }
